admin function for links

+++ gallery/src/main/java/softuniGallery/controller/ArticleController.java
@@ -2,6 +2,7 @@ package softuniGallery.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -92,6 +93,16 @@ public class ArticleController {
             return "redirect:/";
         }
 
+        if (!(SecurityContextHolder.getContext().getAuthentication()
+                instanceof AnonymousAuthenticationToken)) {
+            UserDetails principal = (UserDetails) SecurityContextHolder.getContext()
+                    .getAuthentication().getPrincipal();
+
+            User entityUser = this.userRepository.findByEmail(principal.getUsername());
+
+            model.addAttribute("user", entityUser);
+        }
+
         Article article = this.articleRepository.findOne(id);
 
         model.addAttribute("article", article);
@@ -109,6 +120,11 @@ public class ArticleController {
         }
 
         Article article = this.articleRepository.findOne(id);
+
+        if (!isUserAuthorOrAdmin(article)) {
+            return "redirect:/article/" + id;
+        }
+
         List<Category> categories = this.categoryRepository.findAll();
         String tagString = article.getTags().stream().map(Tag::getName).collect(Collectors.joining(", "));
         model.addAttribute("view", "article/edit");
@@ -164,6 +180,10 @@ public class ArticleController {
 
         Article article = this.articleRepository.findOne(id);
 
+        if (!isUserAuthorOrAdmin(article)) {
+            return "redirect:/article/" + id;
+        }
+
         model.addAttribute("article", article);
         model.addAttribute("view", "article/delete");
 
@@ -179,6 +199,10 @@ public class ArticleController {
         }
 
         Article article = this.articleRepository.findOne(id);
+
+        if (!isUserAuthorOrAdmin(article)) {
+            return "redirect:/article/" + id;
+        }
 
         this.articleRepository.delete(article);
